commit 14

+++ src/main/java/be/vdab/services/FiliaalServiceImpl.java
@@ -5,18 +5,23 @@ import org.springframework.stereotype.Service;
 
 import be.vdab.dao.FiliaalDAO;
 import be.vdab.entities.Filiaal;
+import be.vdab.exceptions.FiliaalHeeftNogWerknemersException;
+import be.vdab.exceptions.FiliaalMetDezeNaamBestaatAlException;
 
 @Service
 public class FiliaalServiceImpl implements FiliaalService {
 	private final FiliaalDAO filiaalDAO;
-	
+
 	@Autowired
 	public FiliaalServiceImpl(FiliaalDAO filiaalDAO) {
 		this.filiaalDAO = filiaalDAO;
 	}
-	
+
 	@Override
 	public void create(Filiaal filiaal) {
+		if (filiaalDAO.findByNaam(filiaal.getNaam()) != null) {
+			throw new FiliaalMetDezeNaamBestaatAlException();
+		}
 		filiaalDAO.create(filiaal);
 	}
 
@@ -27,14 +32,19 @@ public class FiliaalServiceImpl implements FiliaalService {
 
 	@Override
 	public void update(Filiaal filiaal) {
+		Filiaal anderFiliaal = filiaalDAO.findByNaam(filiaal.getNaam());
+		if (anderFiliaal != null && anderFiliaal.getId() != filiaal.getId()) {
+			throw new FiliaalMetDezeNaamBestaatAlException();
+		}
 		filiaalDAO.update(filiaal);
-
 	}
 
 	@Override
 	public void delete(long id) {
+		if (filiaalDAO.findAantalWerknemers(id) != 0) {
+			throw new FiliaalHeeftNogWerknemersException();
+		}
 		filiaalDAO.delete(id);
-
 	}
 
 	@Override
